requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

sentence-transformers==2.2.2
rank-bm25==0.2.2
faiss-cpu==1.7.4
torch==2.1.1
numpy==1.24.3

chromadb==0.4.22
sqlalchemy==2.0.23
sqlite3

pdfplumber==0.10.3
python-docx==0.8.11
python-magic-bin==0.4.14

requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1

structlog==23.2.0
